refactor(tree): lift ShellRun/ShellResult to engine/runtime/shell.ts
Moves the ShellRun and ShellResult type declarations out of
sync.ts into a dedicated runtime/shell.ts module. open-tree-pr.ts
previously imported ShellRun from sync.ts, creating a
circular-ish dependency between two sibling engine primitives.

Gardener has its own locally-defined ShellRun/ShellResult that
carries an env?: NodeJS.ProcessEnv option tree's runner doesn't
need — intentionally not unified here per scope sign-off on #196.

Zero behavior change: sync-golden-snapshot fixtures pass without
UPDATE_SYNC_GOLDEN=1; all 342 tree tests green.

Closes #196.
